+++ dns-integrator-operator/src/charm.py
@@ -72,7 +72,9 @@ class DnsIntegratorCharm(ops.CharmBase):
         if not self.model.unit.is_leader():
             return
         try:
-            self.dns_record.update_relation_data(self._get_dns_record_data())
+            dns_record_data = self._get_dns_record_data()
+            for relation in self.dns_record.relations:
+                self.dns_record.update_relation_data(dns_record_data, relation)
         except ops.model.ModelError as e:
             logger.error("ERROR while updating relation data: %s", e)
             raise